+++ test/secp256k1_key_pair_test.dart
@@ -38,7 +38,7 @@ void main() {
 
       final actual = await keyPair.verify(
         dataToSign,
-        signature: signature,
+        signature,
       );
 
       expect(actual, isTrue);